escape % characters in strings
The problem here is that perhaps sometimes the character IS actually
for formatting - a db value used later for formatting variables into
a string. Also, this means the translator will likely put %% as the
translation, and because this is not really python, it will
actually show '%%' in the translated data, so the %% need to be
translated to %...

A proper fix would have to detect this properly, and/or format things
back on input, or instead of making it appear to be a python file, use
an html intermediary format.