Update structure from review
Moves the test directory inside the app, which is the usual django app
location for them.

Removes settings from conftest, and conftest itself, and puts the
settings.py and urls.py to a test_project directory.

Adds pytest-django config to a pytest.ini to point at the
test settings file