makemessages: qualify vinfile strings as unicode
In Python 2, an io file requires unicode input, and will raise an
exceptino if given plain str objects, so we convert its input strings
accordingly.

I checked that the resulting vinfile produces correct gettext()
statements for both Python 2 and 3, so we should be okay.

This will also work in Python 3 since the u-prefix is merely
informative there.